api/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.6.0
rasterio>=1.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
tqdm>=4.66.0
openpyxl>=3.1.0

# Optional for DEM processing
pydem>=0.1.0

# File handling
python-magic>=0.4.27
aiofiles>=23.2.1
